+++ source/Messaging/source/Messaging/Subscriber/ISubscriber.cs
@@ -12,11 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Google.Protobuf;
-
-namespace Energinet.DataHub.Core.Messaging.Communication.Internal;
+namespace Energinet.DataHub.Core.Messaging.Communication.Subscriber;
 
 /// <summary>
-/// ADR-008:  https://energinet.atlassian.net/wiki/spaces/D3/pages/328957986/ADR+008+-+Integration+events+with+protocol+buffers
+/// Handles incoming messages from the Service Bus.
 /// </summary>
-public record IntegrationEvent(Guid EventIdentification, string EventName, int EventMinorVersion, IMessage Message);
+public interface ISubscriber
+{
+    Task HandleAsync(IntegrationEventServiceBusMessage message);
+}