+++ source/Messaging/documents/release-notes/release-notes.md
@@ -1,5 +1,10 @@
 # Messaging Release notes
 
+## Version 3.1.0
+
+- Support for receiving messages from Service Bus using a background service or Service Bus trigger.
+- Support for publishing messages ad hoc, for when a background service is not suitable.
+
 ## Version 3.0.0
 
 - New implementation for sending messages via Service Bus using a background service.